+++ lib/src/model/video_encoder_configuration.g.dart
@@ -13,12 +13,14 @@ VideoEncoderConfiguration _$VideoEncoderConfigurationFromJson(
       useCount: OnvifUtil.mappedToInt(json['UseCount'] as Map<String, dynamic>),
       encoding: OnvifUtil.nullableMappedToString(
           json['Encoding'] as Map<String, dynamic>?),
-      resolution:
-          Resolution.fromJson(json['Resolution'] as Map<String, dynamic>),
+      resolution: json['Resolution'] == null
+          ? null
+          : Resolution.fromJson(json['Resolution'] as Map<String, dynamic>),
       quality:
           OnvifUtil.mappedToDouble(json['Quality'] as Map<String, dynamic>),
-      rateControl:
-          RateControl.fromJson(json['RateControl'] as Map<String, dynamic>),
+      rateControl: json['RateControl'] == null
+          ? null
+          : RateControl.fromJson(json['RateControl'] as Map<String, dynamic>),
       mpeg4: json['MPEG4'] == null
           ? null
           : Mpeg4.fromJson(json['MPEG4'] as Map<String, dynamic>),
@@ -36,9 +38,9 @@ Map<String, dynamic> _$VideoEncoderConfigurationToJson(
       'Name': instance.name,
       'UseCount': instance.useCount,
       'Encoding': instance.encoding,
-      'Resolution': instance.resolution.toJson(),
+      'Resolution': instance.resolution?.toJson(),
       'Quality': instance.quality,
-      'RateControl': instance.rateControl.toJson(),
+      'RateControl': instance.rateControl?.toJson(),
       'MPEG4': instance.mpeg4?.toJson(),
       'H264': instance.h264?.toJson(),
       'Multicast': instance.multiCast.toJson(),